+++ package.json
@@ -20,9 +20,8 @@
     "react-router-dom": "^6.14.2",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4",
-    "yup": "^1.2.0"
-    "react-scroll": "^1.8.9",
-    "web-vitals": "^2.1.4"
+    "yup": "^1.2.0",
+    "react-scroll": "^1.8.9"
   },
   "scripts": {
     "start": "react-scripts start",
