+++ src/App.js
@@ -5,7 +5,7 @@ import RootLayout from "./Layouts/RootLayout";
 import Home from "./Pages/Home/Home";
 import React from 'react';
 import Category from './Pages/Category/Category';
-import {SignupPage} from './Pages/Signup';
+import {Signup} from './Pages/Signup/Signup';
 
 
 const routes = createBrowserRouter(
@@ -13,7 +13,7 @@ const routes = createBrowserRouter(
         <Route path="/" element={<RootLayout/>}>
             <Route path="/" element={<Home/>}/>
             <Route path='/products/:category' element={<Category/>}/>
-            <Route path='/signup' element={<SignupPage/>}/>
+            <Route path='/signup' element={<Signup/>}/>
         </Route>
     )
 );
